macOS native build support for Apple Silicon added (UE5) (#9730)
* macOS: native Apple Silicon build support (UE5)

* Docs: add macOS build guide

* tuning & docs: add Cook Memory Tuning guidance and update default for 16GB systems

---------

+++ LibCarla/source/carla/Buffer.h
@@ -74,7 +74,7 @@ namespace carla {
         _data(std::make_unique<value_type[]>(size)) {}
 
     /// @copydoc Buffer(size_type)
-    explicit Buffer(uint64_t size)
+    explicit Buffer(size_t size)
       : Buffer([size]() {
           if (size > max_size()) {
             throw_exception(std::invalid_argument("message size too big"));
@@ -93,7 +93,7 @@ namespace carla {
     }
 
     /// @copydoc Buffer(size_type)
-    explicit Buffer(const value_type *data, uint64_t size)
+    explicit Buffer(const value_type *data, size_t size)
       : Buffer(data, [size]() {
           if (size > max_size()) {
             throw_exception(std::invalid_argument("message size too big"));
@@ -322,7 +322,10 @@ namespace carla {
     template <typename T>
     typename std::enable_if<boost::asio::is_const_buffer_sequence<T>::value>::type
     copy_from(size_type offset, const T &source) {
-      reset(boost::asio::buffer_size(source) + offset);
+      // Compute the total size in uint64_t so the checked reset(uint64_t)
+      // overload is used; a size_type cast here would silently truncate
+      // buffer sequences larger than 4 GB.
+      reset(static_cast<uint64_t>(boost::asio::buffer_size(source)) + offset);
       DEBUG_ASSERT(boost::asio::buffer_size(source) == size() - offset);
       DEBUG_ONLY(auto bytes_copied = )
       boost::asio::buffer_copy(buffer() + offset, source);
